Fixed the load balancing lesson
