--- Microservicios con Java/patrones de diseño/load balancer/post-service/src/main/java/com/aldosanchez/post_service/PostServiceApplication.java
@@ -1,15 +1,0 @@
-package com.aldosanchez.post_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class PostServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PostServiceApplication.class, args);
-	}
-
-}
