--- Microservicios con Java/patrones de diseño/load balancer/post-service/src/main/java/com/aldosanchez/post_service/model/Post.java
@@ -1,23 +1,0 @@
-package com.aldosanchez.post_service.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter @Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class Post {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String title;
-    private Long user_id;
-}
