--- Microservicios con Java/patrones de diseño/load balancer/post-service/src/main/java/com/aldosanchez/post_service/repository/IPostRepository.java
@@ -1,15 +1,0 @@
-package com.aldosanchez.post_service.repository;
-
-import com.aldosanchez.post_service.model.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface IPostRepository extends JpaRepository<Post, Long> {
-
-    @Query("SELECT p FROM Post p WHERE p.user_id = :user_id")
-    List<Post> findPostByUserId (Long user_id);
-}
