--- Microservicios con Java/patrones de diseño/load balancer/post-service/src/main/java/com/aldosanchez/post_service/service/IPostService.java
@@ -1,11 +1,0 @@
-package com.aldosanchez.post_service.service;
-
-import com.aldosanchez.post_service.model.Post;
-
-
-import java.util.List;
-
-public interface IPostService {
-
-    public List<Post> getPostByUser(Long user_id);
-}
