--- Microservicios con Java/patrones de diseño/load balancer/post-service/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-server.port=8084
-spring.application.name=post-service
-
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka
-
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/posts_service?serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=
-spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
