--- Microservicios con Java/patrones de diseño/load balancer/post-service/src/test/java/com/aldosanchez/post_service/PostServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.aldosanchez.post_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PostServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
