--- Microservicios con Java/patrones de diseño/load balancer/user-service/src/main/java/com/aldosanchez/user_service/UserServiceApplication.java
@@ -1,17 +1,0 @@
-package com.aldosanchez.user_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-@EnableFeignClients
-public class UserServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserServiceApplication.class, args);
-	}
-
-}
