--- Microservicios con Java/patrones de diseño/load balancer/user-service/src/main/java/com/aldosanchez/user_service/dto/PostDTO.java
@@ -1,17 +1,0 @@
-package com.aldosanchez.user_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter @Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class PostDTO {
-    private Long id;
-    private String title;
-    private Long user_id;
-
-
-}
