--- Microservicios con Java/patrones de diseño/load balancer/user-service/src/main/java/com/aldosanchez/user_service/dto/UserDTO.java
@@ -1,20 +1,0 @@
-package com.aldosanchez.user_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter @Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDTO {
-
-    private Long user_id;
-    private String name;
-    private String lastname;
-    private String cellphone;
-    private List<PostDTO> postList;
-}
