--- Microservicios con Java/patrones de diseño/load balancer/user-service/src/main/java/com/aldosanchez/user_service/repository/IUserRepository.java
@@ -1,9 +1,0 @@
-package com.aldosanchez.user_service.repository;
-
-import com.aldosanchez.user_service.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IUserRepository extends JpaRepository<User,Long> {
-}
