--- Microservicios con Java/patrones de diseño/load balancer/user-service/src/main/java/com/aldosanchez/user_service/service/IUserService.java
@@ -1,8 +1,0 @@
-package com.aldosanchez.user_service.service;
-
-import com.aldosanchez.user_service.dto.UserDTO;
-
-public interface IUserService {
-
-    public UserDTO getUserAndPost(Long user_id);
-}
