--- Microservicios con Java/patrones de diseño/load balancer/user-service/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-server.port=8083
-spring.application.name=user-service
-
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka
-
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/users_service?serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=
-spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
