--- Microservicios con Java/patrones de diseño/load balancer/user-service/src/test/java/com/aldosanchez/user_service/UserServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.aldosanchez.user_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
