+++ Microservicios con Java/patrones de diseño/load-balancer/posts-service/src/main/java/com/aldosanchez/posts_service/PostsServiceApplication.java
@@ -1,0 +1,15 @@
+package com.aldosanchez.posts_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class PostsServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PostsServiceApplication.class, args);
+	}
+
+}
