+++ Microservicios con Java/patrones de diseño/load-balancer/posts-service/src/main/java/com/aldosanchez/posts_service/service/IPostService.java
@@ -1,0 +1,10 @@
+package com.aldosanchez.posts_service.service;
+
+import com.aldosanchez.posts_service.model.Post;
+
+import java.util.List;
+
+public interface IPostService {
+
+    public List<Post> getPostByUser(Long user_id);
+}
