+++ Microservicios con Java/patrones de diseño/load-balancer/posts-service/src/main/resources/application.properties
@@ -1,0 +1,10 @@
+server.port=8082
+spring.application.name=posts-service
+
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka
+
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.url=jdbc:mysql://localhost:3306/posts_service?serverTimezone=UTC
+spring.datasource.username=root
+spring.datasource.password=
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
