+++ Microservicios con Java/patrones de diseño/load-balancer/posts-service/src/test/java/com/aldosanchez/posts_service/PostsServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.aldosanchez.posts_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PostsServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
