+++ Microservicios con Java/patrones de diseño/load-balancer/users-service/src/main/java/com/aldosanchez/users_service/UsersServiceApplication.java
@@ -1,0 +1,17 @@
+package com.aldosanchez.users_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+@EnableFeignClients
+public class UsersServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UsersServiceApplication.class, args);
+	}
+
+}
