+++ Microservicios con Java/patrones de diseño/load-balancer/users-service/src/main/java/com/aldosanchez/users_service/dto/PostDTO.java
@@ -1,0 +1,15 @@
+package com.aldosanchez.users_service.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter @Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class PostDTO {
+    private Long id;
+    private String title;
+    private Long user_id;
+}
