+++ Microservicios con Java/patrones de diseño/load-balancer/users-service/src/main/java/com/aldosanchez/users_service/dto/UserDTO.java
@@ -1,0 +1,21 @@
+package com.aldosanchez.users_service.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserDTO {
+
+    private Long user_id;
+    private String name;
+    protected String lastname;
+    private String cellphone;
+    private List<PostDTO> postsList;
+}
