+++ Microservicios con Java/patrones de diseño/load-balancer/users-service/src/main/java/com/aldosanchez/users_service/model/User.java
@@ -1,0 +1,25 @@
+package com.aldosanchez.users_service.model;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter @Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long user_id;
+    private String name;
+    protected String lastname;
+    private String cellphone;
+
+}
