+++ Microservicios con Java/patrones de diseño/load-balancer/users-service/src/main/java/com/aldosanchez/users_service/repository/IPostsAPI.java
@@ -1,0 +1,16 @@
+package com.aldosanchez.users_service.repository;
+
+import com.aldosanchez.users_service.dto.PostDTO;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+
+import java.util.List;
+
+@FeignClient(name="posts-service")
+public interface IPostsAPI {
+
+    @GetMapping("/posts/{user_id}")
+    public List<PostDTO> getPostByUserId(
+            @PathVariable("user_id") Long user_id);
+}
