+++ Microservicios con Java/patrones de diseño/load-balancer/users-service/src/main/java/com/aldosanchez/users_service/repository/IUserRepository.java
@@ -1,0 +1,9 @@
+package com.aldosanchez.users_service.repository;
+
+import com.aldosanchez.users_service.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IUserRepository extends JpaRepository<User,Long> {
+}
