+++ Microservicios con Java/patrones de diseño/load-balancer/users-service/src/main/java/com/aldosanchez/users_service/service/IUserService.java
@@ -1,0 +1,8 @@
+package com.aldosanchez.users_service.service;
+
+import com.aldosanchez.users_service.dto.UserDTO;
+
+public interface IUserService {
+
+    public UserDTO getUserAndPost(Long user_id);
+}
