+++ Microservicios con Java/patrones de diseño/load-balancer/users-service/src/test/java/com/aldosanchez/users_service/UsersServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.aldosanchez.users_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UsersServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
